Ongoing work for Irix 64-bit mode.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@283 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_energy.c
@@ -574,7 +574,7 @@ mxd_energy_motor_find_home_position( MX_MOTOR *motor )
 
 	MX_ENERGY_MOTOR *energy_motor;
 	MX_RECORD *dependent_motor_record;
-	int direction;
+	long direction;
 	mx_status_type mx_status;
 
 	mx_status = mxd_energy_motor_get_pointers( motor, &energy_motor,
@@ -598,7 +598,7 @@ mxd_energy_motor_constant_velocity_move( MX_MOTOR *motor )
 
 	MX_ENERGY_MOTOR *energy_motor;
 	MX_RECORD *dependent_motor_record;
-	int direction;
+	long direction;
 	mx_status_type mx_status;
 
 	mx_status = mxd_energy_motor_get_pointers( motor, &energy_motor,
