+++ libMx/d_modbus_aio.c
@@ -221,7 +221,7 @@ mxd_modbus_ain_read( MX_ANALOG_INPUT *ainput )
 	if ( mx_status.code != MXE_SUCCESS )
 		return mx_status;
 
-	num_bits = modbus_ainput->num_bits;
+	num_bits = (int) modbus_ainput->num_bits;
 
 	if ( num_bits > 32 ) {
 		return mx_error( MXE_UNSUPPORTED, fname,
@@ -420,7 +420,7 @@ mxd_modbus_aout_write( MX_ANALOG_OUTPUT *aoutput )
 	if ( mx_status.code != MXE_SUCCESS )
 		return mx_status;
 
-	num_bits = modbus_aoutput->num_bits;
+	num_bits = (int) modbus_aoutput->num_bits;
 
 	if ( num_bits > 32 ) {
 		return mx_error( MXE_UNSUPPORTED, fname,
