+++ libMx/i_dsp6001.h
@@ -29,11 +29,11 @@ MX_API mx_status_type mxi_dsp6001_create_record_structures( MX_RECORD *record );
 MX_API mx_status_type mxi_dsp6001_finish_record_initialization(
 							MX_RECORD *record );
 
-MX_API mx_status_type mxi_dsp6001_get_lam_status( MX_CAMAC *crate, int *lam_n);
+MX_API mx_status_type mxi_dsp6001_get_lam_status( MX_CAMAC *crate, long *lam_n);
 MX_API mx_status_type mxi_dsp6001_controller_command( MX_CAMAC *crate,
-								int command );
+								long command );
 MX_API mx_status_type mxi_dsp6001_camac( MX_CAMAC *crate,
-		int slot, int subaddress, int function_code,
+		long slot, long subaddress, long function_code,
 		int32_t *data, int *Q, int *X );
 
 /* Define the data structures used by the DSP 6001/6002 interface code. */
