+++ libMx/i_keithley2000.h
@@ -7,7 +7,7 @@
  *
  *--------------------------------------------------------------------------
  *
- * Copyright 2005 Illinois Institute of Technology
+ * Copyright 2005-2006 Illinois Institute of Technology
  *
  * See the file "LICENSE" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -22,7 +22,7 @@ typedef struct {
 
 	MX_INTERFACE port_interface;
 
-	int last_measurement_type;
+	long last_measurement_type;
 } MX_KEITHLEY2000;
 
 /* 'measurement_type' values. */
